feat: - Starting new architecture file adn directory to upgrade the scalabilty of the project
