+++ web/src/features/courses/components/courseCard/CourseCard.component.tsx
