+++ web/src/features/courses/components/lessonModal/LessonModal.component.tsx
