+++ web/src/features/courses/hooks/useCourseDetailPage.ts
