+++ web/src/features/courses/hooks/useCoursesPage.ts
