+++ web/src/features/courses/pages/CourseDetail.page.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ScreenLayout from "@components/ui/common/ScreenLayout.component";
-import ChapterAccordion from "@components/ChapterAccordion/ChapterAccordion.component";
-import { useCourseDetailScreen } from "@hooks/useCourseDetailPage";
+import ChapterAccordion from "@/features/courses/components/chapterAccordion/ChapterAccordion.component";
+import { useCourseDetailScreen } from "@/features/courses/hooks/useCourseDetailPage";
 import { ArrowLeft } from "lucide-react";
-import "@styles/CourseDetailScreen.css";
+import "@features/courses/styles/CourseDetailScreen.css";
 import miloFoxImage from "/miloBook.webp";
 
 const CHAPTER_EMOJIS = ["📘", "📗", "📙", "📕", "📓", "📔", "📒", "📃"];
