+++ web/src/features/courses/styles/CourseDetailScreen.css
