+++ web/src/features/courses/styles/CoursesScreen.css
