+++ web/src/features/landing/components/ParentFaqCard.component.tsx
