+++ web/src/hooks/index.ts
@@ -1,11 +1,10 @@
 // Hooks d'authentification
-export { useLoginForm } from './useLoginForm';
-export { useRegisterForm } from './useRegisterForm';
+export { useLoginForm } from "./useLoginForm";
+export { useRegisterForm } from "./useRegisterForm";
 
 // Hooks de pages
-export { useHomePage } from './useHomePage';
-export { useProfilePage } from './useProfilePage';
-export { useDashboard } from './useDashboard';
+export { useHomePage } from "./useHomePage";
+export { useDashboard } from "./useDashboard";
 
 // Types exportés
-export type { Mission } from './useHomePage';
+export type { Mission } from "./useHomePage";
