+++ web/tsconfig.json
@@ -17,8 +17,7 @@
       "@constants/*": ["src/constants/*"],
       "@types/*": ["src/types/*"],
       "@hooks/*": ["src/hooks/*"],
-      // "@assets/*": ["assets/*"],
-      // "@services/*": ["src/services/*"],$
+      "@features/*": ["src/features/*"],
       "@styles/*": ["src/styles/*"],
       "@store/*": ["src/store/*"],
       "@api/*": ["src/api/*"]
